% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapSwitch.r
\name{bootstrapSwitch}
\alias{bootstrapSwitch}
\title{bootstrapSwitch}
\usage{
bootstrapSwitch(inputId, label, state = TRUE, width = "auto",
  size = "mini", onColor = "success", offColor = "danger")
}
\arguments{
\item{inputId}{inputId used by Shiny}

\item{label}{The label for our switch}

\item{state}{the state of the component which should either be TRUE or FALSE}

\item{width}{the width of our component}

\item{size}{bootstap size parameter}

\item{onColor}{The color the switch shoudl have in on state}

\item{offColor}{The color the switch shoudl have in off state}
}
\description{
creates a bootstrapSwitch component
}
\details{
initialize a bootstrapSwitch
}
